requirements.ci.txt
argparse
requests==2.32.3
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
